+++ Amethyst/Errors/InvalidOverrideSignatureError.cs
@@ -1,0 +1,9 @@
+using Geode.Errors;
+
+namespace Amethyst.Errors
+{
+    public class InvalidOverrideSignatureError(string name) : GeodeError($"Invalid signature to override \"{name}\"")
+    {
+
+    }
+}