+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 * Allow constants in branches
 * Dynamic function calls
+* Better handling of edgecase struct errors
+* Virtual struct methods